test: split local review module coverage
